requirements.txt
flask
gspread
google-auth
python-dotenv
slack_bolt
APScheduler
gunicorn
